+++ lib/J3/VMCore/JavaArray.h
@@ -53,13 +53,13 @@ public:
   
   static T getElement(const TJavaArray* self, uint32_t i) {
     llvm_gcroot(self, 0);
-    assert(i < self->size);
+    assert((ssize_t)i < self->size);
     return self->elements[i];
   }
 
   static void setElement(TJavaArray* self, T value, uint32_t i) {
     llvm_gcroot(self, 0);
-    assert(i < self->size);
+    assert((ssize_t)i < self->size);
     self->elements[i] = value;
   }
 
@@ -94,7 +94,7 @@ public:
   
   static JavaObject* getElement(const ArrayObject* self, uint32_t i) {
     llvm_gcroot(self, 0);
-    assert(i < self->size);
+    assert((ssize_t)i < self->size);
     return self->elements[i];
   }
